Moved NodeWrapper.h from include/generic to include/
