+++ include/NodeWrapper.h
