+++ src/capra_camera_3d/src/capra_camera_3d.cpp
@@ -1,4 +1,4 @@
-#include "generic/NodeWrapper.h"
+#include "NodeWrapper.h"
 // #include "PACKAGE/MSG TYPE HEADER"
 
 #include <chrono>
@@ -14,7 +14,7 @@ int main(int argc, char** argv)
     // to let 3D camera initialize all its
     // topics
     using namespace std::chrono_literals;
-    std::this_thread::sleep_for(5s);
+    std::this_thread::sleep_for(10s);
 
     /*
     typedef <MSG TYPE HERE> Message;
